perf: avoid huge crypto polyfill
