+++ pages/_app.page.tsx
@@ -6,7 +6,6 @@ import type { AppProps } from "next/app";
 import { useRouter } from "next/router";
 import { Header } from "./header/Header";
 import { GetServerSideProps } from "next";
-import { isAuthenticated } from "./api/auth/authenticated";
 import React from "react";
 import Head from "next/head";
 import dayjs from "dayjs";
@@ -19,16 +18,6 @@ dayjs.extend(relativeTime);
 type HigherProps<P> = { props: P; authenticated: boolean };
 export type ServerProps<P> = GetServerSideProps<HigherProps<P>>;
 export const Authenticated = React.createContext<boolean | null>(null);
-
-export function serverProps<P extends object>(
-  props: GetServerSideProps<P>,
-): ServerProps<P> {
-  return async (context) => {
-    const result = (await props(context)) as any;
-    result.props.authenticated = isAuthenticated(context.req, context.res);
-    return result;
-  };
-}
 
 export default function App({ Component, pageProps }: AppProps) {
   // Force data reload on fast linking:
